+++ packages/editor/web/src/RichContentEditor/Decorators/Resize/createDecorator.js
@@ -193,12 +193,11 @@ export default ({ config, store }) => WrappedComponent =>
         ...elementProps
       } = this.props;
       const componentData = blockProps.getData();
-      const config = componentData.config || {};
-      const { size, alignment } = config;
+      const { size, alignment } = componentData.config;
 
       const {
-        width = componentData.config?.width,
-        height = componentData.config?.height,
+        width = componentData.config.width,
+        height = componentData.config.height,
         hoverPosition,
       } = this.state;
       const { isTop, isLeft, isRight, isBottom } = hoverPosition;